[変更] プレイヤー別のスコア位置調整

+++ Assets/Scripts/View/ScoreView.cs
@@ -6,7 +6,7 @@ public class ScoreView : MonoBehaviour
     [SerializeField]
     private Text Points;
 
-    public void ChangePointsPosition(int x, int y)
+    public void ChangePointsPosition(float x, float y)
     {
         Points.rectTransform.anchoredPosition = new Vector3(x, y, 0);
     }
